requirements.txt
Django==3.2.7
requests==2.26.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0